lab3-fastapi-service/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Vector Store
pinecone-client==3.0.0
chromadb==0.4.22

# LLM and Embeddings
openai==1.10.0
sentence-transformers==2.3.1
langchain==0.1.5
langchain-openai==0.0.3

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
wikipedia-api==0.6.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
flake8==7.0.0

